requirements.txt
faiss-cpu
numpy~=2.0.2
streamlit~=1.45.1
sentence-transformers
langchain~=0.3.25
transformers~=4.51.3
pymupdf
